refactor(server): reorganize HTTP validation and documentation schemas
- Introduce dedicated response schemas for error and success cases
- Separate auth validation schemas into individual files
- Create shared schemas for common fields like email and password
- Update route definitions to use new schema structure
- Remove obsolete schema files
