+++ server/src/infra/http/docs/responses/error/badRequestResponse.ts
@@ -1,0 +1,7 @@
+import z from 'zod'
+
+export const badRequestResponse = z.object({
+  statusCode: z.literal(400),
+  error: z.string(),
+  message: z.string(),
+})
