+++ server/src/infra/http/docs/responses/error/unauthorizedResponse.ts
@@ -1,0 +1,7 @@
+import z from 'zod'
+
+export const unauthorizedResponse = z.object({
+  statusCode: z.literal(401),
+  error: z.string(),
+  message: z.string(),
+})
