+++ server/src/infra/http/routes/auth.routes.ts
@@ -2,8 +2,8 @@ import type { FastifyInstance } from 'fastify'
 import type { ZodTypeProvider } from 'fastify-type-provider-zod'
 import login from '../../../application/usecase/Login'
 import register from '../../../application/usecase/Register'
-import { loginRouteSchema } from '../schemas/LoginRoute'
-import { registerRouteSchema } from '../schemas/RegisterRoute'
+import { loginRouteSchema } from '../docs/routes/auth/loginRoute'
+import { registerRouteSchema } from '../docs/routes/auth/registerRoute'
 
 export async function authRoutes(server: FastifyInstance) {
   server
