+++ server/src/infra/http/validation/auth/loginSchema.ts
@@ -1,0 +1,10 @@
+import { z } from 'zod'
+import { emailSchema } from '../shared/emailSchema'
+import { passwordSchema } from '../shared/passwordSchema'
+
+export const loginSchema = z.object({
+  email: emailSchema,
+  password: passwordSchema,
+})
+
+export type LoginInput = z.infer<typeof loginSchema>
