+++ server/src/infra/http/validation/shared/emailSchema.ts
@@ -1,0 +1,7 @@
+import z from 'zod'
+
+export const emailSchema = z
+  .string()
+  .email('O e-mail digitado deve ser um e-mail válido')
+  .max(255, 'O limite de caracteres do e-mail é 255')
+  .nonempty('O e-mail é obrigatório')
